Refatorando rotas.
