--- backend/routes.js
@@ -1,12 +1,0 @@
-const { Router } = require('express');
-const UserController = require('./src/controllers/UserController');
-
-const routes = Router();
-
-routes.get('/', UserController.firstMessage);
-routes.get('/user/:id', UserController.read);
-routes.post('/user', UserController.create);
-routes.put('/user/:id', UserController.update);
-routes.delete('/user/:id', UserController.delete);
-
-module.exports = routes;
